event_handler/requirements.txt
pydantic
fastapi
starlette
uvicorn
pymongo
python-dotenv
httpx 
azure-storage-blob
apscheduler